requirements.txt
tensorflow
numpy
scipy
matplotlib
gwpy
gwosc
jupyter
scikit-learn
seaborn
pillow
gradio 
pillow
streamlit
gdown